+++ koku/masu/celery/tasks.py
@@ -567,4 +567,4 @@ def get_celery_queue_items(self, queue_name=None, task_name=None):
 @celery_app.task(name="masu.celery.tasks.delayed_summary_polling", queue=GET_REPORT_FILES_QUEUE)
 def delayed_summary_polling(*args, **kwargs):
     """Removes the expired records starting the delayed celery tasks."""
-    DelayedCeleryTasks.remove_expired_records()
+    DelayedCeleryTasks.delayed_summary_polling()
