+++ koku/masu/processor/tasks.py
@@ -61,6 +61,7 @@ from masu.util.gcp.common import deduplicate_reports_for_gcp
 from masu.util.oci.common import deduplicate_reports_for_oci
 from reporting.ingress.models import IngressReports
 from reporting_common.models import CostUsageReportStatus
+from reporting_common.models import DelayedCeleryTasks
 from reporting_common.states import ManifestState
 from reporting_common.states import ManifestStep
 
@@ -109,6 +110,29 @@ QUEUE_LIST = [
 ]
 
 UPDATE_SUMMARY_TABLES_TASK = "masu.processor.tasks.update_summary_tables"
+
+
+def delayed_summarize_current_month(schema_name: str, provider_uuids: list, provider_type: str):
+    """Delay Resummarize provider data for the current month."""
+    queue = UPDATE_SUMMARY_TABLES_QUEUE
+    if is_customer_large(schema_name):
+        queue = UPDATE_SUMMARY_TABLES_QUEUE_XL
+
+    for provider_uuid in provider_uuids:
+        id = DelayedCeleryTasks.create_or_reset_timeout(
+            task_name=UPDATE_SUMMARY_TABLES_TASK,
+            task_args=[schema_name],
+            task_kwargs={
+                "provider_type": provider_type,
+                "provider_uuid": str(provider_uuid),
+                "start_date": str(DateHelper().this_month_start),
+            },
+            provider_uuid=provider_uuid,
+            queue_name=queue,
+        )
+        if schema_name == settings.QE_SCHEMA:
+            # bypass the wait for QE
+            id.delete()
 
 
 def record_all_manifest_files(manifest_id, report_files, tracing_id):
